requirements.txt
# Core
django==4.2.9
python-dotenv==1.0.1
str2bool==1.1
djangorestframework==3.14.0
instaloader==4.10.3
drf-yasg==1.21.7
setuptools==69.1.0

# UI
django-admin-black==1.0.8

# Deployment
whitenoise==6.6.0
gunicorn==21.2.0
django-cors-headers

# psycopg2-binary
# mysqlclient==2.1.1
pymysql

# Hot Reload
django-browser-reload==1.12.1

# PayPal SDK
paypalrestsdk==1.13.3

# Pillow
Pillow==10.2.0

#stripe
stripe==11.1.0